// vue/src/store/generated/kiprasdak/energija/kiprasdak.energija.energija/module/types/energija/smart_meter.d.ts
import { Writer, Reader } from "protobufjs/minimal";
export declare const protobufPackage = "kiprasdak.energija.energija";
export interface SmartMeter {
    index: string;
    production: number;
    consumption: number;
    description: string;
}
export declare const SmartMeter: {
    encode(message: SmartMeter, writer?: Writer): Writer;
    decode(input: Reader | Uint8Array, length?: number): SmartMeter;
    fromJSON(object: any): SmartMeter;
    toJSON(message: SmartMeter): unknown;
    fromPartial(object: DeepPartial<SmartMeter>): SmartMeter;
};
declare type Builtin = Date | Function | Uint8Array | string | number | undefined;
export declare type DeepPartial<T> = T extends Builtin ? T : T extends Array<infer U> ? Array<DeepPartial<U>> : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>> : T extends {} ? {
    [K in keyof T]?: DeepPartial<T[K]>;
} : Partial<T>;
export {};